requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.10.0
blinker==1.9.0
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.2.1
command-not-found==0.3
cppman==0.5.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Flask==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.2
oauthlib==3.2.0
opencv-python==4.11.0.86
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
PyYAML==5.4.1
requests==2.25.1
rsa==4.9.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
systemd-python==234
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.1.3
zipp==1.0.0
